requirements.txt
agency-swarm
pyreadline3
selenium
gradio
python-dotenv